Safari-609.1.11.1/LayoutTests/pointer-lock/pointer-lock-api.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../http/tests/resources/js-test-pre.js"></script>
</head>
<body>
<script>
    description("Basic API existence test for Pointer Lock.")

    shouldBeDefined("document.onpointerlockchange");
    shouldBeDefined("document.onpointerlockerror");
    shouldBeDefined("document.exitPointerLock");
    shouldEvaluateTo("document.pointerLockElement", null);
    element = document.createElement("div");
    shouldBeDefined("element.requestPointerLock");
</script>
<script src="../http/tests/resources/js-test-post.js"></script>
</body>
</html>
